Don't inherit from type defined in MS.ANC.Razor.LanguageServer
