+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/Cohost/CohostSemanticTokensLegendService.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
 using System.ComponentModel.Composition;
-using Microsoft.AspNetCore.Razor.LanguageServer.Semantic;
 using Microsoft.CodeAnalysis.Razor.SemanticTokens;
 using Microsoft.CodeAnalysis.Razor.Workspaces.Protocol;
 
@@ -10,6 +9,7 @@ namespace Microsoft.VisualStudio.LanguageServerClient.Razor.Cohost;
 
 [Export(typeof(ISemanticTokensLegendService))]
 [method: ImportingConstructor]
-internal class CohostSemanticTokensLegendService(IClientCapabilitiesService clientCapabilitiesService) : RazorSemanticTokensLegendService(clientCapabilitiesService)
+internal class CohostSemanticTokensLegendService(IClientCapabilitiesService clientCapabilitiesService)
+    : AbstractRazorSemanticTokensLegendService(clientCapabilitiesService)
 {
 }
